extract interfaces for MQTT client configuration
