+++ mqtt-hivemq/src/main/java/io/micronaut/mqtt/hivemq/v3/config/Mqtt3ClientConfiguration.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2017-2023 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.mqtt.hivemq.v3.config;
+
+import io.micronaut.mqtt.hivemq.config.MqttClientConfiguration;
+
+/**
+ * MQTT v3 Client Configuration.
+ * @author Sergio del Amo
+ * @since 3.0.0
+ */
+public interface Mqtt3ClientConfiguration extends MqttClientConfiguration {
+    boolean isCleanSession();
+}
